add a limit to the amount of notes onscreen
i was gonna hold off on this for a bit but then i noticed nael's revamped spamming mod did this as well, so I eventually implemented it fully, as well as renamed 2 other options.

and then when i was checking the rendered note counter i noticed that the issue had popped back up where sustain notes wouldn't be removed properly. so I fixed that as well.

and THEN i realized that the max rendered thing wasn't working anymore as it tracked the length of both notes and sustainNotes added together. so I implemented a veeeery easy fix for that as well!